Refactoring of utils folders structure
