+++ src/constants/date.js
@@ -1,5 +1,6 @@
 import moment from 'moment';
 
 export const baseFormat = 'YYYY-MM-DD';
+export const baseFormatWithTime = `${baseFormat} HH:mm:ss`;
 
 export const dateNow = moment().format(baseFormat);
