--- src/utils/constants.js
@@ -1,1 +1,0 @@
-export const DATE_RANGE_FILTER_FORMAT = 'YYYY-MM-DDTHH:mm:ss.SSS';
