+++ src/utils/mappers.js
@@ -1,7 +1,5 @@
 import moment from 'moment/moment';
-import { baseFormat } from '../../../src/constants';
-
-window.moment = moment;
+import { baseFormat } from '../constants';
 
 const DATA_TYPES = {
   NUMERIC: 'NUMERIC',
@@ -35,6 +33,7 @@ export const getMappedTableData = (data) => {
   const contentData = data.rows?.map(({ row }) => {
     return row.reduce((acc, item, index) => {
       const column = data.header[index];
+
       // it's required format the dates on FE side
       acc[column.value] =
         column.dataType === DATA_TYPES.DATE_TIME
